+++ src/example/mood/components/happy/types.ts
@@ -1,8 +1,6 @@
 import { Action } from "../../../../library/index.ts";
-import { Mood, MulticastActions } from "../../types.ts";
+import { Mood } from "../../types.ts";
 
 export class Actions {
-  static Multicast = MulticastActions;
-
   static Select = Action<Mood>("Select");
 }